Update Adafruit_FXAS21002C.cpp - line 221 warning

+++ Adafruit_FXAS21002C.cpp
@@ -218,7 +218,7 @@ void Adafruit_FXAS21002C::getSensor(sensor_t *sensor) {
   sensor->type = SENSOR_TYPE_GYROSCOPE;
   sensor->min_delay = 0;
   sensor->max_value = (float)this->_range * SENSORS_DPS_TO_RADS;
-  sensor->min_value = (this->_range * -1.0) * SENSORS_DPS_TO_RADS;
+  sensor->min_value = ((float)this->_range * -1.0) * SENSORS_DPS_TO_RADS;
   sensor->resolution = 0.0F; // TBD
 }
